Refactor hook creation
